Function implemented to check whether logged in User is in fact Owner of Termin

+++ backend/src/main/java/com/hsaugsburg/HRManagementTool/services/CalendarService.java
@@ -2,17 +2,14 @@ package com.hsaugsburg.HRManagementTool.services;
 
 import com.hsaugsburg.HRManagementTool.database.entity.MitarbeiterEntity;
 import com.hsaugsburg.HRManagementTool.database.entity.TerminEntity;
-import com.hsaugsburg.HRManagementTool.database.repository.MitarbeiterRepo;
 import com.hsaugsburg.HRManagementTool.database.repository.TerminRepo;
-import com.hsaugsburg.HRManagementTool.dto.calendar.TerminUpdateDTO;
 import com.hsaugsburg.HRManagementTool.mapper.calendar.CalendarMapper;
-import com.hsaugsburg.HRManagementTool.models.calendar.Termin;
-import com.hsaugsburg.HRManagementTool.models.calendar.TerminUpdate;
+import com.hsaugsburg.HRManagementTool.services.models.calendar.Termin;
+import com.hsaugsburg.HRManagementTool.services.models.calendar.TerminUpdate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.security.core.Authentication;
 
-import java.util.Optional;
 import java.util.Set;
 
 
@@ -36,12 +33,19 @@ public class CalendarService {
 
     }
 
-    public void deleteTermin(int terminId) {
+    public void deleteTermin(int terminId, Authentication authentication) {
+        TerminEntity terminToDeleted =  this.terminRepo.findById(terminId).orElseThrow();
+        if(!isLoggedInUserOwnerOfTermin(terminToDeleted.getErstellerId(),authentication)){
+            throw new RuntimeException("ERROR! Logged in User is not owner of Termin");
+        }
         this.terminRepo.deleteById(terminId);
     }
 
-    public Termin updateTermin(TerminUpdate terminUpdate) {
+    public Termin updateTermin(TerminUpdate terminUpdate,Authentication authentication) {
        Termin terminToBeUpdated= getTermin(terminUpdate.getId());
+       if(!isLoggedInUserOwnerOfTermin(terminToBeUpdated.getErstellerId(),authentication)){
+           throw new RuntimeException("ERROR! Logged in User is not owner of Termin");
+       }
        terminToBeUpdated.update(terminUpdate);
        TerminEntity terminToBeUpdatedEntity = calendarMapper.mapToTerminEntity(terminToBeUpdated);
        Set<MitarbeiterEntity> teilnehmerOfUpdatedTermin = this.mitarbeiterService.getMitarbeiterEntities(terminUpdate.getTeilnehmer());
@@ -51,8 +55,15 @@ public class CalendarService {
 
     }
 
-    public Termin getTermin(final int terminId){
+    private Termin getTermin(final int terminId){
         return this.terminRepo.findById(terminId).map(termin -> calendarMapper.mapToTermin(termin)).orElseThrow();
     }
+
+    private boolean isLoggedInUserOwnerOfTermin(int erstellerID,Authentication authentication){
+        MitarbeiterEntity loggedInUser = this.mitarbeiterService.getMitarbeiterEntity(authentication.getName());
+        return erstellerID == loggedInUser.getId();
+    }
+
+
 
 }
